Bilateral Relations
Remove the "our relation with them / their relation with us"
paradigm in favor of a single global relation between two races.

This feature was requested in 2015.

Effect on game dynamics
=======================

Due to the reduced number of players, sometimes races don't have
any active players. This makes it impossible for other races to
ever interact with that race. For the benefit of those still
playing, this change gives them the ability to make slow progress
towards peace with a defunct race.

Since there are 7 other races, which could all in theory be
choosing to change relations with your race, you are no longer
in direct control over your relations. You can still counter some
of them with your votes, and you still have the trump card: a
war vote, which will allow any active council to prevent peace
with another race if they so choose.

I also consider this to satisfy point 2 in issue #53, since a
bilateral relation means that relations can increase twice as
fast. As such, we will keep the relation change at +/-10.

Effect on code performance
==========================

The relations were actually stored (or at least queried) backwards
from the standpoint of performance. Previously, to get your relations
with a race, you would query that race's relations with _all other_
races (see `Globals::getRaceRelations`). So if you needed to get
your relations with all races, you needed to query 8*8=64 relations!

With bilateral relations, all you will ever need to query are the 8
numbers that represent your race's relations with the other races
(and, equivalently, their relations with your race).

The database structure hasn't changed, so we just need to be aware
that querying `Globals::getRaceRelations` with the current player's
race ID is the preferred usage in all normal scenarios.