Dockerfile: Add git commit data to the docker image
Add enough data from the `.git` directory to be able to store
the commit hash of HEAD. This allows us to tell which version
of the code is running in a docker image.

Ideally we would want this in the image metadata, but I think
Docker prevents this sort of dynamic logic at build time.

We use this method in favor of two other alternatives:

1. Docker Cloud build arg

Docker Cloud provides a `$SOURCE_COMMIT` build arg, but it can
only be passed into the build environment with a custom build
hook. Unfortunately, this causes the Docker Cloud build to not
utilize the layer cache correctly, even though diagnostic
messages indicate that it is pulling/pushing the layer index
for the repo. I don't understand why it does this, and Docker
Cloud support is very limited. I suspect it is because the
default build hook (which is not public) has special magic that
the recommended build command override is missing.

See https://github.com/docker/hub-feedback/issues/600 for details.

2. Custom build script

If we want to forego Docker Cloud altogether, we can write our
own custom script to build images and push them to the Docker
registry. This would give us complete flexibility, but sacrifices
automation and uses personal resources instead of free commercial
services. We may have to do this anyway if Docker Cloud continues
its pattern of deactivating service components.