Add "Accepted" status for feature requests
This status will help organize the (dauntingly long) list of open
feature requests. Many are good ideas that I plan to implement,
but they continue to clutter the list of open requests that still
need to be dealt with. It will also help give players some quicker
feedback about what is and isn't planned for implementation.