SmrGalaxy: optimize getGalaxyContaining
In profiling the Current Sector page, I found that a significant
fraction (roughly 10-20%) of the cost was from constructing an
`SmrGalaxy` for every galaxy in the game (see `getGameGalaxies`)
while checking to determine which galaxy the current sector is in.
This was unnecessary since the `sector` table includes the galaxy
containing each sector, which we query instead.

* Replace a few instances of `getGalaxyContaining` with a new
  function `SmrSector::sectorExists`, which more clearly indicates
  what its purpose is.

* To help with verifying sector existence, add a new exception
  `SectorNotFoundException`, which is thrown in the `SmrSector`
  constructor instead of the base Exception.