SmrForce: single db query for grouped calls
In the style of `SmrPlayer`, allow passing a db query result
into the `SmrForce` constructor, which then skips the db query
inside the constructor. This allows us to then perform a single
query in `getSectorForces` to avoid the cost of performing a
query for _every_ stack of forces in a sector.

* Modify `getForce` and `SmrForce` ctor to take a
  `gameIDOrResultArray` as the first argument.

* Fix a bug with `forceUpdate` that would have prevented
  `getSectorForces` from updating the cache correctly (but this
  option was not in use, so the bug was only theoretical).