SmrPlayer: add npc field
We often have to get the display name for a list of players.
Since we append "[NPC]" to the display name for NPC players,
we have to check if each player is an NPC, which entails:

1) querying the associated account to get the account login
2) cross-checking the login with the `npc_logins` table

These two queries triple the already large expense of displaying
lists of players.

By adding a boolean `npc` field to the `player` table, we
can entirely avoid this additional cost. (This could also be
achieved with a compound query, but the boolean is the simplest
and most efficient solution.)