SmrShip: only change hardware that has changed
There are up to 11 pieces of hardware at the moment, but often
only one of them is changing at a time (e.g. shields/forces).
However, we previously would update every value if any single
hardware amount changed, resulting in up to 10 unnecessary db
queries.

Because of `set{Old}Hardware`, it is easy to keep track of
which hardware has changed, and only perform the db query to
update those that have changed.