Better error message when a force stack is full
If we know that the only reason no forces are added is because
the stack can't take any more of the forces the player tried to
drop, then say so in the error message (rather than "You want to
add/remove 0 forces?", which can be confusing for someone who
doesn't know there's a cap on force stack size).

This was suggested by RCK in 2012.