shop_goods.php: allow supply to ajax update
The middle panel ajax is disabled on the main port page (because of
the input elements, see `Template::checkDisableAJAX`).

However, we often find ourselves refreshing the port page to see
the supply/demand incrementing, so we add the `ajax` class and a
unique id to its column to enable ajax on that specific element.

NOTE: if using this to watch for supply to increase enough to
complete a trade, the player will have to remember that the auto-fill
value for the trade will not change (since auto-updating that with
ajax would obliterate any attempt to fill in the form manually).