SmrShip: update cargo display immediately
When ship cargo changes, the data in memory is not consistent with
the data when it is loaded from the database. This means that the
cargo will be displayed one way when it first changes, and then on
the next page it will be displayed differently.

I found at least two scenarios where this occurs:

1. When you have more than one good and then sell or dump one of
   the goods, it will then display with zero amount (e.g. "Ore: 0").
   To fix this, we remove the field from the cargo property when
   we delete it from the database so that it never shows up at all.

2. When you already have one good in your cargo and you buy a
   second kind of good, if the second good has a lower ID than the
   first, it will display after the first good, and then switch
   on the next page. If we sort the cargo property when we first
   add the new good, then it will always display in numerical ID
   order.