planet_financial.php: ajax update financial page
The planet balance and bond can be ajax updated by wrapping them
in a span with a unique id.

The balance is always displayed, so we can span just the number.
However, the bond message is omitted if there is no bond, so we
need to the span around the entire sentence (and outside the
conditional).