Default.css: fix shifting page center
For pages that are too long, the browser adds a vertical scrollbar.
Because the page content is centered, the added scrollbar shifts
the content every so slightly to the left.

By making the page width only as wide as the viewport width _minus_
the scrollbar, adding the scrollbar no longer perturbs the page
content alignment.

See https://stackoverflow.com/a/25981719.