newsletter_send_processing.php: set unlimited timeout
Give PHP an unlimited amount of time to do the mass mailing,
since it may take a long time depending on the number of accounts.

This setting will be ignored if PHP is compiled with
`--enable-safe-mode`. Regardless, the web server or browser may
decide to terminate the connection before the task is done.