account_edit.php: fix searching with no account ID
Fix a bug introduced in cd34fdd0625, which resulted because
`$_REQUEST['account_id']` is _always_ specified due to the way
that the form is set up. I thought the default value would be 0,
but it will always be the empty string if not specified. Then
this empty string caused a `$db->escapeNumber` to fail.

We now explciitly set `$account_id = 0` if it is not specified.