Allow using PHP 7+
Add PHP versions up to 7.2.X in the composer.json file and switch
to the latest PHP 7.2 version in Dockerfile.

Going through the list of backwards incompatible changes from 5.6
to 7.2, I don't see anything that affects the SMR codebase.

NOTE: As of PHP 7.1, the `rand()` and `array_rand()` functions use
a better random number generator, and should be considered usable.