SmrPlayer: remove hard-coded pod speed
When a player is killed, their post-death ship can have a different
speed than their pre-death ship, so turns need to be modified.
Having the speed of the escape pod hard-coded in the conversion
causes two problems:

1. We can't safely change the speed of the escape pod.
2. We can't safely allow a ship other than the escape pod
   to be given upon death.

Querying the post-death ship's speed instead solves these problems.

In particular, this fixes a bug where newbie players lose extra turns
when they die, because the Newbie Merchant Vessel (the ship newbies
receive post-death instead of the Escape Pod) is speed 9 and the
Escape Pod is speed 7.