Remove game version 1.2 code
The 1.2 code has serious backwards incompatibilities that would make
restoring functionality too overwhelming. Furthermore, we do not
expect to ever use 1.2, so there is no point in maintaining it.

Keeping the 1.2 code only makes it more difficult to maintain the
non-1.2 code. Since it will not be modified from its current state,
we will remove the code with the understanding that we can always
recover it if needed.