forces_drop_processing.php: fix vanishing forces
Fixes a bug where forces dropped would vanish from ship if you
encountered an error during processing.

To reproduce the error, you would drop cds/scouts and more mines
than you have, or cds and more scouts than you have. Your ship
would update the cds/scouts or cds, but the force stack would not
have its expire time updated, and so they would immediately
disappear.

To fix the error, we simply perform all sanity checks before making
any changes to the ships. Therefore, we ensure that if any changes
are made, the entire script is processed as expected.

A note is added to explain why we don't update the expire
time automatically every time a change is made. In addition to
that note, there is a specious `$forces->update()` call in the
method that updates the expire time, and so we want to avoid the
extra db calls that would incur.