SmrSession: remove LIMIT 1 from deletion
One possible solution to (or mitigation of) the infrequent mysql error:

Duplicate entry '<hash>' for key 'PRIMARY'

is to remove the `LIMIT 1` qualifier from the deletion that directly
precedes the INSERT that is causing this error.

The reason this may help is because certain incompletely understood
race conditions might result in an account having two active sessions.
Then when this deletion comes along, it removes one but not both of
these sessions. It is unclear why the primary keys would be identical.

The `active_session` table is small (due to pruning inactive sessions),
so having to process the entire table when creating a new session is
not going to impact performance at all.