SmrSession: allow active session for each game
Add a new method `SmrSession::updateGame`, which changes the
`game_id`, but also ensures that active sessions are unique for
each (account_id, game_id) pair.

This allows players to have multiple games open for the same
account (though in different browsers, due to the nature of the
browser cookies).

This is also an important step towards integrating multis into
the same account as the main.