Integrate multis into main accounts
The purpose of this is twofold:
* Allow players to navigate their main/multi accounts without
  having to log in and out.
* Synchronize things between the two accounts when it doesn't
  make sense to have two values (i.e. e-mail address, user
  preferences, bans, etc.).

Add a new table `account_link_login`, which associates a `login_id`
(the new primary key replacing `account_id` in the `account` table)
with an `account_id` (the primary key of `account_link_login`).
Each "login" (`account`) can be associated with multiple "accounts"
(`account_link_login`).

This is a little confusing, since ideally the `account` table would
be renamed `login` (or similar), but we are probably a bit too heavily
invested in the original infrastructure to completely change it now.

We provide a way on the "Play Game" page to switch to/create multis,
including an option to link an existing account as a multi.

Adds two helper methods to SmrAccount:
* getLinkedAccountList - finds all accounts associated with login
* getLinkedAccount - returns an account if it is associated with login