Install NBBC with Composer
Use Composer so that we do not need to include the source code
in the smr repo. Update from 1.4.5 to 2.x, which is maintained
by a new group -- the reason for the major version change is
primarily for Composer and PHP7+ support, which breaks backwards
compatibility with 1.x.

NOTE: Page had manually disabled escaping HTML in the included
source code. The newer version we're now using via Composer has
an `setEscapeContent` method, which we set to `false` to skip
escaping HTML. This allows things like the game news and messages
to display as expected (though it can be dangerous to do this in
theory if you're not careful about user-generated messages).