ip_view_results.php: modernize code
This code doesn't use any objects and has extremely old design
patterns. It also emits many warnings and sometimes causes 502
Bad Gateway errors (perhaps due to the overwhelming number of
warnings?).

* Use objects and their methods where it makes sense from a
  DB performance standpoint to do so (e.g. `SmrAccount` and
  `isDisabled`).

* Use `implode` to join lists of strings together.

* Remove the special case for `Account tagged for deletion`
  closing reasons (which isn't a closing reason that exists).

* Since `host` is stored in `account_has_ip`, don't recompute
  it. It's only for informational purposes anyway.

* Minor fixes:
    - Remove an extra column
    - Replace an erroneous `ip` with `$ip`
    - Fix copy-pasted code for result description

There is still a lot of code duplication here, but for now we're
simply trying to fix the bugs and make the code decipherable.