Allow header to overlay Galaxy Map
When we are scrolling around the Galaxy Map for large galaxies,
it is a pain to scroll up to the top left corner to change the
galaxy or center on a sector. By having the header in a fixed
position on the page instead, it is always available for use.

Originally, I wanted the header to have a solid background that
would overlap (and obscure) the galaxy map, and then a little
jQuery button that would minimize the header. I didn't know how
to implement this, so I settled for a transparent background,
but no ability to minimize the header. This means the galaxy
map is always slightly obscured, and the header is a bit harder
to read when it overlaps the galaxy map, but I think it is
better than obscuring a significant portion of the galaxy map.