current_sector.php: fix mini combat display warnings
Pass `AlreadyDestroyed => false` to the included templates for
minimal combat display. Since `AlreadyDestroyed` is only true
for the trader doing the attacking when there is no combat, we
don't get a combat log and so we know that the minimal display
will not be triggered.

This logic is perhaps a bit convoluted, but would probably need
a refactor to make the logic more automatic.