games_previous.php: fix start/end dates
For compatibility games, the database query was extracting the
start/end dates in a date format rather than seconds since epoch.
We were then trying to pass those dates through the PHP `date`
function, which was not valid. It would emit a warning and then
display the date as 1/1/1970.

Now we simply select the start/end dates in seconds since epoch
and convert them to date format in PHP like we do everywhere else.