docker-compose.yml: switch from links to depends_on
The `links` option is a legacy option that may eventually be
removed entirely. We switch to `depends_on` so that we are
still expressing the service dependencies.

Note that since `links` allows a `SERVICE:ALIAS` syntax, and
`depends_on` does not, we need to tell the `phpmyadmin` service
that the mysql container is called `mysql` and not `db`
(the latter being the default container name for the phpmyadmin
image).