PortKillMessage.inc: fix undefined index KillerCredits
We no longer award credits to the player that destroys the port.
Instead, it is awarded to the looter. As a result, we no longer
need the message in `PortKillMessage.inc` about credits, but we
keep it there just in case it's needed in the future. To avoid
the PHP warning, we make the message contingent on the existence
of the `KillerCredits` index.