Plotter: fix findReversiblePathFromX
Fix typos in `Plotter::findReverisblePathFromX` when it was added in
1adb4c1e4d62. These typos allowed players to plot to any location
in the universe, even if they hadn't scouted it.

I've also removed the `isset` workarounds for the default arguments.
It's possible the warning this circumvented was changed in PHP7, but
more likely is that this was a manifestation of the typos.

Thanks to RedRocket for reporting the bug.