SmrPlanet: make planet attack news more reliable
Add a new news entry for each planet attack. Previously an attack
would be omitted if _any_ attack was made recently (even if it was
a different planet). Now we use planet owner to determine uniqueness
of planet attacks.

Increase the time before a new news entry is added for the _same_
planet from 10 mins to 30 mins. This helps avoid duplicate entries
now that we aren't worried about missing other planet attack news.