SectorMap.inc: make unexplored linked sectors more obvious
Previously in the Local Map, the only thing differentiating
explored and unexplored linked sectors was the color of the sector
number (green vs. yellow).

This was really hard to distinguish, so now unexplored linked
sectors will have the explored linked sector color bordering
a black background.

Some CSS changes were required to implement this in the simplest
way I could think of. We remove the `{visited,unvisited}` classes
from the Local Map (they are still used in the Current Sector),
and instead use the information in `{current,connect,normal}Seclm{u,}`
to identify whether or not the sector is explored. Note that the
trailing `u` means unexplored.