Remove references from class method arguments
Arguments should only be passed by reference into a function if
the function intends to modify the original variable. It should
not be used to try to improve performance (it won't: see #317).

Here we remove improper references from class method arguments,
along with a few other internal references that I missed earlier.