Make script runtime more accurate
The script runtime was computed at the end of `do_voodoo`, but before
`$template->display`. As a result, the displayed runtime only included
the cost of processing the `engine` files. However, this entirely
excludes the cost of processing the `template` files, which is the
majority of the cost for some files (e.g. Local and Galaxy maps).

Ideally we would compute the script runtime at the end of `do_voodoo`
just before `exit` is called, but unfortunately there is no way to
then display this number since the html document tags are closed at
the end of `$template->display`.

Since `copyright.inc` is processed (nearly) last on the page, it is a
significant improvement to compute the script runtime inline, so we
do that now . This is still not ideal, but it only excludes a few
milliseconds of processing that occurs afterwards.