Simplify adding warps in universe generation
Added a summary table of all the warps for each galaxy.
The table also provides convenient links to switch the
galaxy that is being edited.

Note: since looping through every sector in the universe is very
expensive, we switch to querying the `warp` table directly. This
way we only have a handful of queries rather than thousands.

* Add more detailed description of how the warp creator works.
* Add a sanity check limit of 10 warps per galaxy pair (in case,
  e.g., someone thinks they're supposed to put sector ID's there).