SmrShip: refactor update methods
The `SmrShip::update` method was called on every page, and it would
make a mysql query (on the `player` table) even if the `shipTypeID`
hadn't changed.

This method was also unnecessary because `SmrShip::setShipTypeID`
calls `SmrPlayer::setShipTypeID`, which ensures that the `player`
table is updated correctly.

First order of business is to remove the direct `player` table
query in `SmrShip::update`. Second is to refactor the various
udpate methods so that they are all called in `SmrShip::update`.
This is mostly unnecessary since `SmrShip::saveShips` is called
on every page, but we make this self-consistent update method
available for consistency with other classes.