SmrPlayer: fix e-mail notifications from non-player accounts
The e-mail notifications for received player messages did not
correctly handle non-player senders. We would be calling
`SmrPlayer::getPlayer` for an account that doesn't exist, which
would throw a database error.

To fix this, we have to instead call `getMessagePlayer`, which
handles non-player senders correctly (this is what we do to
parse senders when viewing/sending messages in-game).