Move warp into sector table
Due to the way we construct `SmrSector`, we end up querying the
`warp` table for every single sector. This essentially doubles the
cost of the `SmrSector` ctor.

On the other hand, adding an extra `warp` field to the `sector`
table adds a negligible cost to the `sector` database query, so we
do this instead.