Silence ircbot ping message by default
The server ping (which occurs roughly every minute) prints a message
that becomes extremely verbose when the ircbot has been running for
a long time. To avoid overloading the logs, we silence this message
by default. An optional config variable `IRC_BOT_VERBOSE_PING` can
be sent to `true` to recover this diagnostic message.