.dockerignore: remove .git directory
Allow the `.git` directory to be sent to the docker daemon.

This is needed for:
* Getting the commit hash as metadata
* The Git enricher for the API docs

To avoid copying the entire `.git` directory into the image
when we copy the SMR source code, we need to explicitly list
the directories we want to copy. This has the added benefit
of omitting all the other non-`.git` contents that we didn't
really want in the image anyway.

(NOTE: it is unfortunate that the Dockerfile syntax for COPY
does not include the ability to copy multiple directories in
a single command without flattening the directory structure.)