Auto-update page to current sector when kicked from planet
In `create_error`, an exception is thrown if `USING_AJAX=true`.
This is because the ajax page refreshes don't really know how
to handle when a different page needs to be displayed without
the user clicking a button.

The most important place where this occurs is when a player is
kicked from a planet (by the owner or in a bust). In the past,
this would cause ajax to stop updating due to the exception.
We could instead forward to the current sector page, but this
would display a disjointed page because a) the left panel links
don't have span id's for updating and so you'd still see the
"Planet Main" link, and b) the middle panel won't update
correctly because we already have ajax-enabled child elements,
which disables the full middle panel update.

Instead what we will do is use `Template::addJavascriptForAjax`
(which needs to be made public so it can be used in the `engine`
files as well as the `template` files) to force the page to
load the current sector, circumventing all the ajax issues.

However, we still need to display _some_ page, since the javascript
won't trigger until after the page loads. Our options are:
1) Display the planet page for maximum continuity, but then if
   the player isn't using javascript, they will be able to
   effectively stay on the planet even when kicked off.
2) Return immediately after adding the javascript (which would
   need to be done both in the `engine` and `template` pages),
   but then a blank middle panel will display (which also isn't
   a great thing to display for non-javascript users).
3) Forward to `current_sector.php` after adding the javascript.
   This only requires us to modify the `engine` files, and
   displays only a marginally broken ajax-updated page.

We also continue to call `create_error` as before whenever
`USING_AJAX=false`, since that will correctly forward to the
current sector and display the error message.

Since this logic is somewhat fragile and is needed for all planet
display pages (not processing pages, since those require an
explicit click by the user), this is a good opportunity to refactor
all the common code out of these pages into `planet.inc`.