loader.php: fix include ordering issue
In 2f705a49e007c, an include using `get_file_loc` was moved above
the part of the code where `$overrideGameID` was defined.
This caused all non-Default games to break, because it loaded the
wrong `Smr*` classes.

We have to be very careful about the ordering of includes here,
and there is now a comment warning about this.