Dockerfile: enable Zend OPcache
Closes #297.

At long last, the mystery is solved. The reason the live server was
~5x faster when it was using the host system's PHP (rather than Docker)
was because it had the Zend OPcache enabled. This PHP extension caches
the compiled PHP and allows it to be reused at runtime. Since we have
some fairly large libraries, loading and parsing each script was a
significant expense.

Simply installing the `opcache` PHP extension in the Dockerfile
dramatically improves performance.