preferences.php: reorganize CSS preferences
Display changes:

* Merge the "Enable Default CSS", "Template", and "Colour Scheme"
  options into a single "CSS Template" option, which lists all
  the possible Template/Colour Scheme variants as well as "None".

  The "None" option is the equivalent of Enable Default CSS == No.
  It is now listed as a separate "CSS Template" option because,
  when disabled, it doesn't matter to the user what the
  Template/Colour Scheme are.

* Rename "Change CSS Link" to "Add CSS Link", and reword the help
  text to make it clearer what this option does. Also removed the
  link to the non-existent CSS file example.

These changes were needed for the following reasons:

a) It wasn't clear how all these knobs interacted.
b) Fewer knobs is easier for the user in general.
c) It was annoying to switch between templates because you often
   needed to change the template, save, then go back and change
   the colour scheme.

The storage of the CSS preference and the logic for how those
preferences are used to render the page is unchanged. This is
_only_ a simplification for the sake of a clearer user interface.