Streamline course plot processing
Because we deep copied the `$path` in `course_plot_processing.php`
and `course_plot_nearest_processing.php` _before_ we performed
the final path processing, we duplicated this work in
`course_plot_result.php`.

If we instead do the common processing in `course_plot_results.php`
_only_, then we avoid both the duplicated code/processing _and_
the cost of the deep copy.

Since `course_plot_results.php` is sometimes a display page, we do
the processing first so that we don't do the display work if we
end up forwarding to `current_sector.php`.

Also:
* Convert `course_plot_results.php` to a template.