planet_main.php: remove duplicate id attributes
The `<img>` tags for the planet hardware had IDs that were
duplicating the ship IDs for `cds` and `shields` (and also had
a duplicate `turret` ID that should have been `armour`). This
resulted in the following warning, e.g.:

> DOMDocument::loadHTML(): ID shields already defined in Entity

To fix these, we simply remove the `id` attributes for these
images, since they are not needed.