SmrPlanet: change timer for planet bust news
Wait 1 hour instead of 30 minutes before generating additional
news entries for an attack on the same planet (with the same owner).

NOTE: if the planet owner changes, another attack at any time will
trigger a news entry.

This was changed because planet busts on big planets can sometimes
last up to an hour. But we don't want the cooldown to be so large
that a second attack isn't announced (let's say an alliance is
attacked once, and then they stop for an hour and come back -- this
should generate two news entries).