composer.json: switch to phpquery dev-master
We need some PHP7 fixes that have been added to phpquery in the
5 years since v0.9.6 was released.

In particular, this fixes the error in phpQueryObject.php:3124

> Error: Call to undefined function split()

I believe this is only an issue if you enable debugging in
phpquery with `phpQuery::$debug = 1`, but it being broken
makes debugging other phpquery issues more difficult.

Specify a hash to get a fixed revision on `dev-master` to prevent
a future push to `dev-master` from silently breaking our code.
(Thanks to Page for the suggestion!)