RightPanelShip.inc: fix nesting rules violation
We used the `<noscript>` tags around opening and closing `<a>` tags
separately. Since this results in open elements when the `<noscript>`
tag is called, it violates nesting rules.

While this still works correctly in most browsers, it breaks HTML
validators in non-recoverable ways and, perhaps more importantly,
causes `DOMDocument::loadHTML` to emit the warning:

> Unexpected end tag : a

While we don't expect many (if any) people to have JavaScript disabled,
we do not yet want to make JavaScript required for gameplay.

The following elements are affected:

1. Show/Hide weapons in the right panel
2. Ship weapon reordering

To fix this, we remove the `<noscript>` tags entirely, thus displaying
the `<a>` elements regardless of whether JS is disabled or not. Then,
we add a `click` action to the link so that, when JS is enabled, it
nullifies the `href` attribute, thus effectively disabling it as a link.

NOTE: We currently disable further click actions to avoid spamming,
so we need to add a case to that routine to prevent disabling clicks
in this case, where the `<a>` element is not a "real" link
(i.e. `href="javascript:void(0);"`).

This lets the pages retain their existing JS-enabled behavior, with
the only addition being a nullified link wrapping the element, which
is an acceptable sacrifice to make to fix the nesting rules violation.